user_proj_example: remove example project
This project is useful for demonstrating how to create projects.
However, its existence causes risk for any external projects that assume
`user_proj_example` is the name of this project.

Remove any references to `user_proj_example` in order to prevent it from
interfering with the real files in this repo.

Signed-off-by: Sean Cross <sean@xobs.io>